Refactor NumericTextBox to support Value binding and improved UX

+++ src/backend/VoltStream.Application/Features/Supplies/Commands/CreateSupplyCommand.cs
@@ -13,7 +13,7 @@ public record CreateSupplyCommand(
     string CategoryName,
     long ProductId,
     string ProductName,
-    decimal RollCount,
+    int RollCount,
     string Unit,
     decimal LengthPerRoll,
     decimal TotalLength,
