+++ src/frontend/ApiServices/Models/Responses/SupplyResponse.cs
@@ -4,7 +4,7 @@ public record SupplyResponse
 {
     public long Id { get; set; }
     public DateTimeOffset Date { get; set; }
-    public decimal RollCount { get; set; }
+    public int RollCount { get; set; }
     public decimal LengthPerRoll { get; set; }
     public decimal TotalLength { get; set; }
     public decimal UnitPrice { get; set; }
